+++ routes/navlist.js
@@ -34,6 +34,7 @@ module.exports = function (req) {
         'list': [
           { 'href': '/admin', 'key': 'admin', 'label': 'Admin', 'title': 'Admin page' },
           { 'href': '/admin/users', 'key': 'adminusers', 'label': 'Users', 'title': 'User administration' },
+          { 'href': '/admin/periods', 'key': 'adminperiods', 'label': 'Periods', 'title': 'Periods administration' },
           { 'key': 'divider' },
           { 'href': '/admin/feedback', 'key': 'adminfeedback', 'label': 'Feedback', 'title': 'Feedback administration' },
           { 'key': 'divider' },